+++ src/lib/form/useOnSubmit.ts
@@ -40,7 +40,7 @@ export type EnhancedOnSubmit<TValues> = (
     }
 ) => SubmissionErrors | Promise<SubmissionErrors> | void
 
-const defaultNavigateTo: GetToFunction = ({
+export const defaultNavigateTo: GetToFunction = ({
     section,
     submitAction = 'saveAndExit',
     responseData,